move conversion util functions into wrapper files
