+++ gym/dev_wrappers/tf_wrapper.py
@@ -1,9 +1,10 @@
 from typing import Any, Dict, Union, Tuple
 
+from jax._src import dlpack as jax_dlpack
+from jax.interpreters.xla import DeviceArray
+
 from gym import Env, Wrapper
 from gym.error import DependencyNotInstalled
-
-from gym.utils.conversions import jax_to_tf, tf_to_jax
 
 try:
     import tensorflow as tf
@@ -11,6 +12,28 @@ except ImportError:
     raise DependencyNotInstalled(
         "tensorflow is not installed, run `pip install tensorflow`"
     )
+
+
+def tf_to_jax(arr: tf.Tensor):
+    # TODO: Verify this captures vector envs and converts everything from tensorflow to jax
+    return jax_dlpack.from_dlpack(tf.experimental.dlpack.to_dlpack(arr))
+
+
+def jax_to_tf(arr):
+    # TODO: Verify this captures vector envs and converts everything from jax to tensorflow
+    return tf.experimental.dlpack.from_dlpack(jax_dlpack.to_dlpack(arr))
+
+
+def tf_dict_to_jax(
+     value: Dict[str, Union[tf.Tensor, Any]]
+) -> Dict[str, Union[DeviceArray, Any]]:
+    return type(value)(**{k: tf_to_jax(v) for k, v in value.items()})
+
+
+def jax_dict_to_tf(
+    value: Dict[str, Union[DeviceArray, Any]]
+) -> Dict[str, Union[tf.Tensor, Any]]:
+    return type(value)(**{k: tf_to_jax(v) for k, v in value.items()})
 
 
 class JaxToTFV0(Wrapper):
