+++ gym/dev_wrappers/torch_wrapper.py
@@ -8,20 +8,54 @@ https://github.com/google/brax/blob/9d6b7ced2a13da0d074b5e9fbd3aad8311e26997/bra
 Under the Apache 2.0 license. Copyright is held by the authors
 """
 
-from typing import Dict, Optional, Tuple, Union
+from typing import Any, Dict, Optional, Tuple, Union
+
+from jax._src import dlpack as jax_dlpack
+from jax.interpreters.xla import DeviceArray
 
 from gym import Env, Wrapper
 from gym.error import DependencyNotInstalled
 
-from gym.utils.conversions import jax_to_torch, jax_dict_to_torch, torch_to_jax
-
 try:
     import torch
+    from torch.utils import dlpack as torch_dlpack
 except ImportError:
     raise DependencyNotInstalled("torch is not installed, run `pip install torch`")
 
 
 Device = Union[str, torch.device]
+
+def torch_to_jax(value: torch.Tensor) -> DeviceArray:
+    """Converts a PyTorch Tensor into a Jax DeviceArray."""
+    tensor = torch_dlpack.to_dlpack(value)
+    tensor = jax_dlpack.from_dlpack(tensor)
+    return tensor
+
+
+def jax_to_torch(value: DeviceArray, device: Device = None) -> torch.Tensor:
+    """Converts a Jax DeviceArray into a PyTorch Tensor."""
+    dlpack = jax_dlpack.to_dlpack(value.astype("float32"))
+    tensor = torch_dlpack.from_dlpack(dlpack)
+    if device:
+        return tensor.to(device=device)
+    else:
+        return tensor
+
+
+def torch_dict_to_jax(
+    value: Dict[str, Union[torch.Tensor, Any]]
+) -> Dict[str, Union[DeviceArray, Any]]:
+    """Converts a dictionary of PyTorch Tensors into a Dictionary of Jax DeviceArrays."""
+    return type(value)(**{k: torch_to_jax(v) for k, v in value.items()})
+
+
+def jax_dict_to_torch(
+     value: Dict[str, Union[DeviceArray, Any]], device: Device = None
+) -> Dict[str, Union[torch.Tensor, Any]]:
+    """Converts a dictionary of Jax DeviceArrays into a Dictionary of PyTorch Tensors."""
+    return type(value)(
+        **{k: jax_to_torch(v, device) for k, v in value.items()}
+    )
 
 
 class JaxToTorchV0(Wrapper):
